- Fixed bug with not spawning task graph communicator when no other graph exists - Added simple CUDA test to test suite

+++ src/htgs/core/memory/CudaMemoryManager.hpp
@@ -46,7 +46,7 @@ class CudaMemoryManager: public MemoryManager<T> {
    */
   CudaMemoryManager(std::string name,
                     CUcontext *contexts,
-                    int memoryPoolSize,
+                    size_t memoryPoolSize,
                     std::shared_ptr<IMemoryAllocator <T>> memoryAllocator,
                     MMType type) : MemoryManager<T>(name, memoryPoolSize, memoryAllocator, type) {
     this->contexts = contexts;
